requirements.txt
anyascii==0.3.1
asgiref==3.6.0
beautifulsoup4==4.8.2
blessed==1.20.0
boto3==1.26.80
botocore==1.29.84
captcha==0.4
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.2
defusedxml==0.7.1
Django==3.2.18
django-allauth==0.45.0
django-cors-headers==3.14.0
django-debug-toolbar==3.2.4
django-extensions==3.2.1
django-filter==21.1
django-modelcluster==5.3
django-recaptcha==3.0.0
django-storages==1.13.2
django-taggit==2.1.0
django-treebeard==4.6.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
gunicorn==20.1.0
html5lib==1.1
idna==3.4
jmespath==1.0.1
l18n==2021.3
mailchimp-marketing==3.0.80
oauthlib==3.2.2
openpyxl==3.1.1
paypal-checkout-serversdk==1.0.1
paypalhttp==1.0.1
Pillow==9.4.0
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.0.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2020.5
requests==2.28.2
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
tablib==3.3.0
telepath==0.3
urllib3==1.26.14
wagtail==2.16.3
wagtail-django-recaptcha==1.0
wcwidth==0.2.6
webencodings==0.5.1
whitenoise==6.4.0
Willow==1.4.1
xlrd==2.0.1
XlsxWriter==3.0.8
xlwt==1.3.0
